api_inference/requirements.txt
fastapi
uvicorn[standard]
redis
motor
python-dotenv
transformers
peft
torch
scikit-learn
scipy
numpy
joblib
